refactor(schemas): migrate Phase 18 endpoint schemas — 4 files, 16 classes (#6042)
- alertmanager_webhook.py: migrate AlertAnnotations, AlertLabels, AlertInstance,
  AlertManagerWebhook → schemas_system.py (also normalize Python 3.10 union
  syntax `str | None` → `Optional[str]` for Pydantic v2 compatibility)
- analytics_agents.py: migrate AgentMetricsResponse, TaskRecordResponse,
  TrackTaskRequest, CompleteTaskRequest → schemas_analytics.py
- analytics_behavior.py: migrate TrackEventRequest, FeatureMetricsResponse,
  UserJourneyResponse, EngagementMetricsResponse → schemas_analytics.py
- integration_database.py: migrate DatabaseConnectionRequest, QueryRequest→
  DBIntegrationQueryRequest, MongoQueryRequest, DatabaseListRequest → schemas_workflows.py

Renames to avoid generic-name collisions:
- QueryRequest → DBIntegrationQueryRequest

Cleanup: removed unused DataResponse, datetime, AlertAnnotations/AlertLabels
imports (some classes were preserved in schemas_system.py for completeness but
not used at endpoint).
